Moved chem feats and load to separate folders
