+++ openpharmacophore/__init__.py
@@ -6,15 +6,23 @@ A library to work with pharmacophores.
 # Handle versioneer
 from ._version import get_versions
 
+
+import pyunitwizard as puw
+
+puw.configure.load_library(['pint', "openmm.unit"])
+puw.configure.set_default_form('pint')
+puw.configure.set_standard_units(['angstroms', 'ps', 'K', 'mole', 'amu', 'e',
+                                 'kJ/mol', 'kJ/(mol*nm**2)', 'N', 'degrees'])
+
+
 # Add imports here
-from ._pyunitwizard import puw
 from .point.pharmacophoric_point import PharmacophoricPoint, distance_between_pharmacophoric_points
 from .pharmacophore.pharmacophore import Pharmacophore
 from .pharmacophore.ligand_receptor.pl_complex import PLComplex
 from .pharmacophore.ligand_based.ligand_based import LigandBasedPharmacophore
 from .pharmacophore.ligand_receptor.ligand_receptor import LigandReceptorPharmacophore
-from .pharmacophore.load_from_file import load_from_file
-from .pharmacophore.load import load
+from .load.load_from_file import load_from_file
+from .load.load import load
 from .screening import VirtualScreening
