+++ openpharmacophore/load/load_from_file.py
