+++ tests/unit_tests/pharmacophore/ligand_receptor/test_pl_complex.py
@@ -2,7 +2,7 @@ import mdtraj as mdt
 import numpy as np
 import openpharmacophore.pharmacophore.ligand_receptor.exceptions as exc
 from openpharmacophore import PLComplex
-from openpharmacophore.pharmacophore.chem_feats import smarts_ligand
+from openpharmacophore.chem_feats.chem_feats import smarts_ligand
 import pyunitwizard as puw
 import pytest
 from copy import deepcopy
